backend/requirements.txt
Django==2.2
django-cors-headers==3.5.0
djangorestframework==3.11.1
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
